+++ crypto/src/crypto/digests/LongDigest.cs
@@ -133,11 +133,11 @@ namespace Org.BouncyCastle.Crypto.Digests
             //
             // add the pad bytes.
             //
-            Update((byte)128);
+            Update(128);
 
             while (xBufOff != 0)
             {
-                Update((byte)0);
+                Update(0);
             }
 
             ProcessLength(lowBitLength, hiBitLength);
